Python improvements: add exception handling, logging, fix style.

+++ NeighborlyAPI/deleteAdvertisement/__init__.py
@@ -1,5 +1,9 @@
+import logging
+import os
+
 import azure.functions as func
 import pymongo
+import pymongo.errors
 from bson.objectid import ObjectId
 
 
@@ -9,18 +13,27 @@ def main(req: func.HttpRequest) -> func.HttpResponse:
 
     if id:
         try:
-            url = "mongodb://cosmos-db-account-es81:9ViJPEBPVrmgwFfQacsKvX2FRJJiNP6ijT8Yv5RqCKgIgZNzFpg0ZaEulmafCPD3LoXTrsMixbQVACDbLrsGBQ==@cosmos-db-account-es81.mongo.cosmos.azure.com:10255/?ssl=true&replicaSet=globaldb&retrywrites=false&maxIdleTimeMS=120000&appName=@cosmos-db-account-es81@"
-            client = pymongo.MongoClient(url)
-            database = client['dbes81']
-            collection = database['advertisements']
+            uri = os.getenv('MyDbConnection')
+            if uri is None:
+                logging.error("DB connection string must be set in MyDbConnection env var.")
+                return func.HttpResponse("Database connection string is not set.", status_code=500)
 
+            client = pymongo.MongoClient(uri)
+            db = client.dbes81
+            collection = db.advertisements
+
+            # query = {'_id': id}
             query = {'_id': ObjectId(id)}
             result = collection.delete_one(query)
+            logging.info(f"Deletion from {db} result: {result}")
             return func.HttpResponse("")
 
-        except:
-            print("could not connect to mongodb")
-            return func.HttpResponse("could not connect to mongodb", status_code=500)
+        except pymongo.errors.ConnectionFailure as e:
+            logging.error(e)
+            return func.HttpResponse("Database connection error.", status_code=500)
+        except Exception as e:
+            logging.error(e)
+            return func.HttpResponse("An error occurred.", status_code=500)
 
     else:
         return func.HttpResponse("Please pass an id in the query string",
